+++ src/test/java/org/opensearch/knn/plugin/action/RestKNNStatsHandlerIT.java
@@ -48,6 +48,7 @@ import static org.opensearch.knn.common.KNNConstants.METHOD_IVF;
 import static org.opensearch.knn.common.KNNConstants.METHOD_PARAMETER_NLIST;
 import static org.opensearch.knn.common.KNNConstants.METHOD_PARAMETER_SPACE_TYPE;
 import static org.opensearch.knn.common.KNNConstants.MODEL_ID;
+import static org.opensearch.knn.common.KNNConstants.MODEL_INDEX_NAME;
 import static org.opensearch.knn.common.KNNConstants.NAME;
 import static org.opensearch.knn.common.KNNConstants.NMSLIB_NAME;
 import static org.opensearch.knn.common.KNNConstants.PARAMETERS;
@@ -341,20 +342,23 @@ public class RestKNNStatsHandlerIT extends KNNRestTestCase {
     public void testModelIndexHealthMetricsStats() throws Exception {
         // Create request that filters only model index
         String modelIndexStatusName = StatNames.MODEL_INDEX_STATUS.getName();
+        // index can be created in one of previous tests, and as we do not delete it each test the check below became optional
+        if (!systemIndexExists(MODEL_INDEX_NAME)) {
 
-        Response response = getKnnStats(Collections.emptyList(), Arrays.asList(modelIndexStatusName));
-        String responseBody = EntityUtils.toString(response.getEntity());
-        Map<String, Object> statsMap = createParser(XContentType.JSON.xContent(), responseBody).map();
+            final Response response = getKnnStats(Collections.emptyList(), Arrays.asList(modelIndexStatusName));
+            final String responseBody = EntityUtils.toString(response.getEntity());
+            final Map<String, Object> statsMap = createParser(XContentType.JSON.xContent(), responseBody).map();
 
-        // Check that model health status is null since model index is not created to system yet
-        assertNull(statsMap.get(StatNames.MODEL_INDEX_STATUS.getName()));
+            // Check that model health status is null since model index is not created to system yet
+            assertNull(statsMap.get(StatNames.MODEL_INDEX_STATUS.getName()));
+        }
 
         createModelSystemIndex();
 
-        response = getKnnStats(Collections.emptyList(), Arrays.asList(modelIndexStatusName));
+        Response response = getKnnStats(Collections.emptyList(), Arrays.asList(modelIndexStatusName));
 
-        responseBody = EntityUtils.toString(response.getEntity());
-        statsMap = createParser(XContentType.JSON.xContent(), responseBody).map();
+        final String responseBody = EntityUtils.toString(response.getEntity());
+        final Map<String, Object> statsMap = createParser(XContentType.JSON.xContent(), responseBody).map();
 
         // Check that model health status is not null
         assertNotNull(statsMap.get(modelIndexStatusName));
@@ -414,7 +418,7 @@ public class RestKNNStatsHandlerIT extends KNNRestTestCase {
 
         createKnnIndex(TEST_INDEX, modelIndexMapping(FIELD_NAME, TEST_MODEL_ID));
 
-        addKNNDocs(TEST_INDEX, FIELD_NAME, DIMENSION, DOC_ID, NUM_DOCS);
+        addKNNDocs(TEST_INDEX, FIELD_NAME, DIMENSION, DOC_ID, NUM_DOCS, () -> adminClient());
 
         final Response response = getKnnStats(Collections.emptyList(), Collections.emptyList());
         final String responseBody = EntityUtils.toString(response.getEntity());
@@ -476,10 +480,10 @@ public class RestKNNStatsHandlerIT extends KNNRestTestCase {
     // Useful settings when debugging to prevent timeouts
     @Override
     protected Settings restClientSettings() {
+        final Settings.Builder builder = noStrictDeprecationModeSettingsBuilder();
         if (isDebuggingTest || isDebuggingRemoteCluster) {
-            return Settings.builder().put(CLIENT_SOCKET_TIMEOUT, TimeValue.timeValueMinutes(10)).build();
-        } else {
-            return super.restClientSettings();
+            builder.put(CLIENT_SOCKET_TIMEOUT, TimeValue.timeValueMinutes(10));
         }
+        return builder.build();
     }
 }