+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 * Add CHANGELOG ([#800](https://github.com/opensearch-project/k-NN/pull/800))
 * Bump byte-buddy version from 1.12.22 to 1.14.2 ([#804](https://github.com/opensearch-project/k-NN/pull/804))
 * Bump numpy version from 1.22.x to 1.24.2 ([#811](https://github.com/opensearch-project/k-NN/pull/811))
+* Add support for integ tests on secured cluster ([#827](https://github.com/opensearch-project/k-NN/pull/827))
 ### Documentation
 ### Maintenance
 ### Refactoring